+++ src/runtime/interpreter.h
@@ -66,6 +66,16 @@ typedef struct {
   size_t node_count;
 } graphion_runtime_hypergraph_value;
 
+typedef struct {
+  int64_t id;
+  const graphion_runtime_graph_value *graph;
+} graphion_runtime_graph_node_value;
+
+typedef struct {
+  const graphion_runtime_graph_value *graph;
+  const graphion_runtime_graph_edge *edge;
+} graphion_runtime_graph_edge_value;
+
 typedef enum {
   GRAPHION_VALUE_NONE = 0,
   GRAPHION_VALUE_INT = 1,
@@ -73,7 +83,9 @@ typedef enum {
   GRAPHION_VALUE_BOOL = 3,
   GRAPHION_VALUE_STRING = 4,
   GRAPHION_VALUE_GRAPH = 5,
-  GRAPHION_VALUE_HYPERGRAPH = 6
+  GRAPHION_VALUE_HYPERGRAPH = 6,
+  GRAPHION_VALUE_GRAPH_NODE = 7,
+  GRAPHION_VALUE_GRAPH_EDGE = 8
 } graphion_runtime_value_kind;
 
 typedef struct {
@@ -84,6 +96,8 @@ typedef struct {
   char string_value[GRAPHION_RUNTIME_STRING_MAX];
   graphion_runtime_graph_value *graph_value;
   graphion_runtime_hypergraph_value *hypergraph_value;
+  graphion_runtime_graph_node_value graph_node_value;
+  graphion_runtime_graph_edge_value graph_edge_value;
 } graphion_runtime_value;
 
 typedef struct {